refactor: 상수 분리 파일명 수정
