+++ constants/emotions.ts
@@ -1,0 +1,17 @@
+export const EMOTION_MAP: { [key: number]: string } = {
+  1: '행복',
+  2: '설렘',
+  3: '희망',
+  4: '사랑',
+  5: '평온',
+  6: '벅참',
+  7: '감동',
+  8: '쾌감',
+  9: '신남',
+  10: '용기',
+  11: '센치함',
+  12: '그리움',
+  13: '위로',
+  14: '슬픔',
+  15: '비장',
+};
