+++ hooks/useEmotions.ts
@@ -1,22 +1,5 @@
 import { useState } from 'react';
-
-const emotionMap: { [key: number]: string } = {
-  1: '행복',
-  2: '설렘',
-  3: '희망',
-  4: '사랑',
-  5: '평온',
-  6: '벅참',
-  7: '감동',
-  8: '쾌감',
-  9: '신남',
-  10: '용기',
-  11: '센치함',
-  12: '그리움',
-  13: '위로',
-  14: '슬픔',
-  15: '비장',
-};
+import { EMOTION_MAP } from 'constants/emotions';
 
 // 임시 명세서 (이번꺼는 아직 작성이 안되어 있어서 우선 왓두유씽 ver.1을 참고했습니다. 추후 업데이트 되면 변경하겠습니다)
 const rawEmotionData = {
@@ -38,7 +21,7 @@ const useEmotions = () => {
     emotionCounts[item.content] = item.num; // ex) { 1: 6, 10: 1 }
   });
 
-  const emotions = Object.entries(emotionMap).map(([id, label]) => ({
+  const emotions = Object.entries(EMOTION_MAP).map(([id, label]) => ({
     id: Number(id),
     label,
     count: emotionCounts[Number(id)],
